x-select: use change() to handle changes from DOM
Ember views automatically call `.change()` on DOM change events, so
we can use that hook instead of adding a change handler in `didInsertElement`
and tearing it down in `willDestroyElement`.

This is now the only place we need to call `_updateValueSingle` or
`_updateValueMultiple`. We don't need to do that when the content changes
from upstream since (in Glimmer) the component will render the difference
automatically. Calling those in an observer on `content.@each` was causing
upstream writes to happen unnecessarily.