[feature] add contextual components support
Contextual components provide a nice syntax for encapsulating components
that would not be much use elsewhere. `x-option` is such a component; it
is not so useful without `x-select`. Indeed, we even assert if there is
no parent `x-select` component.

Because we want to provide a great experience for our users who don't
yet have the time to upgrade Ember, I implemented this using ember-cli's
`treeForAddonTemplates` hook. If the consuming app has a version of
ember that is less than `2.3.0`, the original template is provided to
the app tree. Otherwise, we can take advantage of contextual components
and provide a template that yields a hash with `x-option` for users
to use. I stole this from Ember CLI Version Checker's ["Real World Example"](https://github.com/rwjblue/ember-cli-version-checker/blob/0040430d799c783731301773aeeb32b76f936585/README.md#real-world-example)
in their README.