fix the one-way binding attribute.
This separates the `read` operation from the `write` operation which
turns out to be a very handy thing. It turns out that reading a multi
select value and a single select value is the only difference.

So upon an update, we read the next value (which depends on the
selection type) and then, regardless of whethere it is multi or not, we
`set` the value if it is a two-way binding.