UA sniff for bad MS Edge bug (required selects must change value to be valid --glimmer PR) (#172)
* Add forced value-setting as Edge workaround

This is a hackish compensation for a bug in Edge:
https://developer.microsoft.com/en-us/microsoft-edge/platform/issues/8794503/

Without this, a select with the required flag is marked as invalid
until you change the value. You can change the value yourself and
then change it back to the default, but even that workaround fails
if the select only has one value.

I’m not happy with this and will seek other solutions, but I think
this is an okay temporary fix.

* Hide Edge value bug behind UA sniff

Since this is a nasty browser bug I feel a lot better about hiding it
behind a UA sniff to make sure we're in an edge browser.

What's happening is you can't submit a `<select>` that's `required` in
MS Edge without changing the value first.