Remove check for contextual components (#203)
As of x-select 3.x we're only supporting the use of contextual
components which means we can remove the logic around serving old
templates to older ember versions.

**Why does this fix the Travis CI build?**

This is tricky! There's a bug in `ember-cli-version-checker` which
returns `false` for `components/ember#release` ember versions (which
`ember-try` uses). Since that returns false, no options were being
served in the selects.