Upgrade to Ember 3.4 (x-select v4.0) (#224)
* Attempt to upgrade to ember 3.4

This needs a whole lot of work. We need to remove jquery from
everything.

* Update test suite to use Interactors

This uncovered some bugs and also more refactoring is needed

* Update interactor, fill in tests, remove jquery & `sendAction`

* Add attribute integration tests & remove default `tabindex`

`tabindex` does not need to be set to 0 from the ember component since
that's its default in the DOM. Only apply the attribute when it's
different than 0.

* Remove the rest of jQuery from tests

* Single quotes with prettier

* Unskip tests

* Refactor to new test API

* Remove `fs-branding` addon

That addon was a premature abstraction. No need for it to live in a
different project entirely.

* Update `markdown-code-highlighting` addon

* Remove `embadge` badge from README

Looks like the service is now dead

* Update everything to new angle bracket syntax

* Remove jQuery from everything

* First pass at writing new test helper README docs

* Update CHANGELOG to reflect the changes here

* Update dummy app documentation and write 3.x to 4.x migration guide